src/third_party/blink/web_tests/http/tests/fetch/chromium/data-saver.html

<html>
<title>Tests for 'Save-Data' header.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script> internals.setSaveDataEnabled(false); </script>

<iframe srcdoc="
<!DOCTYPE html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script>

var CHECK_PATH = './resources/check-save-data-header.php';
var METHODS = ['GET', 'POST', 'PUT'];
var REQUESTS =
    METHODS.map(method => new Request(CHECK_PATH, {method: method}));

internals.setSaveDataEnabled(false);

window.top.promise_test(t => {
  internals.setSaveDataEnabled(false);
  return Promise.all(REQUESTS.map(request => fetch(request.clone())))
    .then(responses => Promise.all(responses.map(response => response.text())))
    .then(texts => {
        for (var i = 0; i < METHODS.length; ++i) {
          assert_equals(
              texts[i], 'No Save-Data',
              'Save-Data header should not be sent when disabled. method: ' +
              METHODS[i]);
        }
      });
}, 'fetch() from document with Data-Saver disabled.');


window.top.promise_test(t => {
  internals.setSaveDataEnabled(false);
  var worker =
      new Worker('./resources/data-saver-worker.php?dedicated-disabled');
  return new Promise(resolve =>
                     worker.addEventListener('message', resolve))
     .then(msg => {
         var result = msg.data;
         assert_equals(
             result['worker_script_header'], 'No Save-Data',
             'Save-Data header should not be sent for worker script when ' +
             'disabled.');
         for (var i = 0; i < METHODS.length; ++i) {
           assert_equals(
               result[METHODS[i]], 'No Save-Data',
               'Save-Data header should not be sent when disabled. method: ' +
               METHODS[i]);
         }
       });
}, 'fetch() from dedicated worker with Data-Saver disabled.');

// A test for shared workers is implemented as a browser test because shared
// worker script is requested from the browser process when NetworkService is
// enabled, and internals.setSaveDataEnable() doesn't work in that case.

</script>
"></iframe>

</html>
